requirements.txt
autokeras~=1.1.0
numpy~=1.26.2
openai~=1.6.1
scikit-learn~=1.4.0
tensorflow~=2.15.0
python-dotenv~=1.0.0